Finished implementing Data class, added comprehensive unit test, fixed some bugs, added some service layer functionality, added locking mechanism and concurrency tests

+++ src/suncertify/business/ContractorServices.java
@@ -9,5 +9,5 @@ public interface ContractorServices {
 
 	public void createBooking(Contractor contractor) throws Exception;
 
-	public Map<Long, Contractor> find(ContractorPK primaryKey) throws Exception;
+	public Map<Integer, Contractor> find(ContractorPK primaryKey) throws Exception;
 }
